+++ src/node.h
@@ -6,57 +6,59 @@
 struct path;
 struct cost;
 
-class node {
+class node
+{
 private:
-    long int id;
+    long long int id;
     double lattitude;
     double longitude;
 
-    static long int node_count;
+    static long long int node_count;
 
-    std::unordered_map<long int, path> neighbours;
+    std::unordered_map<long long int, path> neighbours;
 
     /// @brief Key: Graph ID | Value: Heuristic value of this node for the graph
-    std::unordered_map<int, shared_ptr<cost>> heuristics;          
+    std::unordered_map<int, shared_ptr<cost>> heuristics;
 
     // friends
     friend class graph;
-    friend double euclidean_distance(node* a, node* b);
-    friend double euclidean_distance(const long int first_node_id, const long int second_node_id);
+    friend double euclidean_distance(node *a, node *b);
+    friend double euclidean_distance(const long long int first_node_id, const long long int second_node_id);
 
     node();
     node(double lattitude_val, double longitude_val);
-    node(long int id_value, double lattitude_val, double longitude_val);
-    
-    path* calculate_path(const long int neighbour_id);
+    node(long long int id_value, double lattitude_val, double longitude_val);
+
+    path *calculate_path(const long long int neighbour_id);
 
 public:
-    void print_neighbours ();
+    void print_neighbours();
 
-    bool add_neighbour (long int neighbour_id);
+    bool add_neighbour(long long int neighbour_id);
 
-    bool add_neighbour (long int neighbour_id, double distance);
+    bool add_neighbour(long long int neighbour_id, double distance);
 
-    bool add_neighbour(long int neighbour_id, const path &path_to_neighbour);
+    bool add_neighbour(long long int neighbour_id, const path &path_to_neighbour);
 
-    void print_node ();
+    void print_node();
 
-    bool add_heuristic (const int graph_id, const long int destination_node_id);
+    bool add_heuristic(const int graph_id, const long long int destination_node_id);
 
-    shared_ptr<cost> get_heuristic (const int graph_id);
+    shared_ptr<cost> get_heuristic(const int graph_id);
 
-    std::vector<long int> get_neighbour_ids ();
+    std::vector<long long int> get_neighbour_ids();
 };
 
-struct path {
+struct path
+{
     double distance = 0.0;
     double time_walk = 0.0;
     double time_bus = 0.0;
     double time_metro = 0.0;
     double time_drive = 0.0;
-    int popularity = 0;                                                         // Range [0, 10]
+    int popularity = 0; // Range [0, 10]
 };
 
-double euclidean_distance(node* a, node* b);
+double euclidean_distance(node *a, node *b);
 
-double euclidean_distance(const long int first_node_id, const long int second_node_id);
+double euclidean_distance(const long long int first_node_id, const long long int second_node_id);
